// src/app/components/dynamic-form/enums/input-mask-type/text-input-format.enum.ts
// <input type="button">
// <input type="checkbox">
// <input type="color">
// <input type="date">
// <input type="datetime-local">
// <input type="file">
// <input type="month">
// <input type="radio">
// <input type="range">

// <input type="email">
// <input type="hidden">
// <input type="image">
// <input type="number">
// <input type="password">

// <input type="reset">
// <input type="search">
// <input type="submit">
// <input type="tel">
// <input type="text">
// <input type="time">
// <input type="url">
// <input type="week">

// TODO: test
export enum TextInputFormat {
    cnpj = 'cnpj',
    cpf = 'cpf',
    integer = 'integer',
    float = 'float',
    date = 'date',
    email = 'email',
    percentage = 'percentage',
    phone = 'phone',
    time = 'time',
    zipCode = 'zip-code',
}